+++ hermes_shm/benchmark/data_structure/atomic.cc
@@ -167,7 +167,7 @@ class AtomicInstructionTestSuite {
     if (test_count == 0) {
       TestOutputHeader();
     }
-    int nthreads = omp_get_num_threads();
+    size_t nthreads = omp_get_num_threads();
     ss << test_name << ","
        << atomic_type_ << ","
        << sizeof(T) << ","
@@ -187,7 +187,7 @@ class AtomicInstructionTestSuite {
  * */
 template<typename AtomicT, typename T, std::memory_order MemoryOrder>
 void TestAtomicInstructionsPerThread() {
-  int count = (1<<20);
+  size_t count = (1<<20);
   AtomicInstructionTestSuite<AtomicT, T, MemoryOrder>().AtomicIncrement(count);
   AtomicInstructionTestSuite<AtomicT, T, MemoryOrder>().AtomicFetchAdd(count);
   AtomicInstructionTestSuite<AtomicT, T, MemoryOrder>().AtomicAssign(count);
