+++ hermes_shm/test/unit/data_structures/lock/lock.cc
@@ -19,9 +19,9 @@ using hshm::Mutex;
 using hshm::RwLock;
 
 void MutexTest() {
-  int nthreads = 8;
-  int loop_count = 10000;
-  int count = 0;
+  size_t nthreads = 8;
+  size_t loop_count = 10000;
+  size_t count = 0;
   Mutex lock;
 
   HERMES_THREAD_MANAGER->GetThreadStatic();
@@ -31,7 +31,7 @@ void MutexTest() {
   {
     // Support parallel write
 #pragma omp barrier
-    for (int i = 0; i < loop_count; ++i) {
+    for (size_t i = 0; i < loop_count; ++i) {
       lock.Lock();
       count += 1;
       lock.Unlock();
@@ -42,22 +42,22 @@ void MutexTest() {
   }
 }
 
-void barrier_for_reads(std::vector<int> &tid_start, int left) {
-  int count;
+void barrier_for_reads(std::vector<int> &tid_start, size_t left) {
+  size_t count;
   do {
     count = 0;
-    for (int i = 0; i < left; ++i) {
+    for (size_t i = 0; i < left; ++i) {
       count += tid_start[i];
     }
   } while (count < left);
 }
 
 void RwLockTest() {
-  int nthreads = 8;
-  int left = nthreads / 2;
+  size_t nthreads = 8;
+  size_t left = nthreads / 2;
   std::vector<int> tid_start(left, 0);
-  int loop_count = 100000;
-  int count = 0;
+  size_t loop_count = 100000;
+  size_t count = 0;
   RwLock lock;
 
   HERMES_THREAD_MANAGER->GetThreadStatic();
@@ -71,7 +71,7 @@ void RwLockTest() {
 
     // Support parallel write
 #pragma omp barrier
-    for (int i = 0; i < loop_count; ++i) {
+    for (size_t i = 0; i < loop_count; ++i) {
       lock.WriteLock();
       count += 1;
       lock.WriteUnlock();
@@ -83,19 +83,19 @@ void RwLockTest() {
     // Support for parallel read and write
 #pragma omp barrier
 
-    int cur_count = count;
-    if (tid < left) {
+    size_t cur_count = count;
+    if ((size_t)tid < left) {
       lock.ReadLock();
       tid_start[tid] = 1;
       barrier_for_reads(tid_start, left);
-      for (int i = 0; i < loop_count; ++i) {
+      for (size_t i = 0; i < loop_count; ++i) {
         REQUIRE(count == cur_count);
       }
       lock.ReadUnlock();
     } else {
       barrier_for_reads(tid_start, left);
       lock.WriteLock();
-      for (int i = 0; i < loop_count; ++i) {
+      for (size_t i = 0; i < loop_count; ++i) {
         count += 1;
       }
       lock.WriteUnlock();
